synching local changes for paper
